test(bb_vm): Adding URL tests

+++ bb_data/models/models_stripe.py
@@ -87,7 +87,7 @@ class ResourceTimeTracking(models.Model):
         total_3070 = int(self.minutes_3070/60) * float(self.rate_3070)
         total_3090 = int(self.minutes_3090/60) * float(self.rate_3090)
 
-        return "%0.2f" % round((total_3070+total_3090), 2) #pylint: disable=consider-using-f-string
+        return float("%0.2f" % round((total_3070+total_3090), 2)) #pylint: disable=consider-using-f-string
 
     class Meta:
         verbose_name_plural = "Resource Time Tracking"
